server: Distribute log functions to user & people/index
